Move CommentStatement to a separate file
