+++ core/src/main/scala/org/bykn/bosatsu/CommentStatement.scala
@@ -1,0 +1,48 @@
+package org.bykn.bosatsu
+
+import cats.data.NonEmptyList
+import cats.implicits._
+import fastparse.all._
+import org.typelevel.paiges.{ Doc, Document }
+
+import org.bykn.fastparse_cats.StringInstances._
+
+/**
+ * Represents a commented thing. Commented[A] would probably
+ * be a better name
+ */
+final case class CommentStatement[T](message: NonEmptyList[String], on: T)
+
+object CommentStatement {
+  type Maybe[T] = Either[T, CommentStatement[T]]
+
+  implicit def document[T: Document]: Document[CommentStatement[T]] =
+    Document.instance[CommentStatement[T]] { comment =>
+      import comment._
+      val block = Doc.intercalate(Doc.line, message.toList.map { mes => Doc.char('#') + Doc.text(mes) })
+      block + Doc.line + Document[T].document(on)
+    }
+
+  implicit def maybeDocument[T: Document]: Document[Maybe[T]] = {
+    val docT = Document[T]
+    Document.instance[Maybe[T]] {
+      case Left(t) => docT.document(t)
+      case Right(c) => document[T](docT).document(c)
+    }
+  }
+
+  /** on should make sure indent is matching
+   * this is to allow a P[Unit] that does nothing for testing or other applications
+   */
+  def parser[T](onP: Parser.Indy[T]): Parser.Indy[CommentStatement[T]] = Parser.Indy { indent =>
+    val commentBlock: P[NonEmptyList[String]] =
+      P(commentPart ~ ("\n" ~ indent ~ commentPart).rep() ~ ("\n" | End))
+        .map { case (c1, cs) => NonEmptyList(c1, cs.toList) }
+
+    P(commentBlock ~ onP(indent))
+      .map { case (m, on) => CommentStatement(m, on) }
+  }
+
+  val commentPart: P[String] = P("#" ~/ CharsWhile(_ != '\n').?.!)
+}
+
